+++ src/vertexbuffer/vertexarray.cpp
@@ -41,16 +41,16 @@ VertexArray::VertexArray(const VertexDeclaration& decl, int size) : tag(0), _dec
 
 VertexArray::Description VertexArray::generateDescription() const
 {
-	uint32_t dataSize = 0;
-	uint32_t offset = 0;
+	size_t dataSize = 0;
+	size_t offset = 0;
 
 	Description desc;
 	desc.declaration = VertexDeclaration(_decl.interleaved());
 
 	for (auto& chunk : _chunks)
 	{
-		int t_stride = _decl.interleaved() ? static_cast<int>(_decl.dataSize()) : 0;
-		uint32_t t_offset = _decl.interleaved() ? offset : dataSize;
+		uint32_t t_stride = static_cast<uint32_t>(_decl.interleaved() ? _decl.dataSize() : 0);
+		uint32_t t_offset = static_cast<uint32_t>(_decl.interleaved() ? offset : dataSize);
 		desc.declaration.push_back(VertexElement(chunk->usage(), chunk->type(), t_stride, t_offset));
 		dataSize += chunk->dataSize();
 		offset += chunk->typeSize();
